+++ test-files/Example.kt
@@ -1,0 +1,8 @@
+// This is a test Kotlin file
+/**
+ * Main function for testing
+ */
+fun main() {
+    // Print hello message
+    println("Hello World")
+}