+++ test-files/Example.scala
@@ -1,0 +1,8 @@
+// This is a test Scala file
+/**
+ * Main object for testing
+ */
+object Example extends App {
+  // Print hello message
+  println("Hello World")
+}